enable IDEA CE 2023.2.5 support
