+++ src/main/java/org/openjpa/ide/idea/PersistentState.java
@@ -1,21 +1,29 @@
 package org.openjpa.ide.idea;
 
+import com.intellij.openapi.components.PersistentStateComponent;
+import com.intellij.openapi.components.Service;
+import com.intellij.openapi.components.Storage;
+import com.intellij.openapi.project.Project;
+import com.intellij.util.xmlb.XmlSerializerUtil;
+import com.intellij.util.xmlb.annotations.XCollection;
+import org.jetbrains.annotations.NotNull;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.LinkedHashSet;
 
-import com.intellij.util.xmlb.annotations.AbstractCollection;
-
-import org.openjpa.ide.idea.integration.EnhancerSupport;
-
 /**
  * Holds plugin's persistent state.
  */
-public class PersistentState { // has to be public (for IDEA configuration access)
+@Service(Service.Level.PROJECT)
+@com.intellij.openapi.components.State(name = "OpenJpaConfiguration",
+        storages = {@Storage(value = "openjpa-plugin.xml")})
+public final class PersistentState implements PersistentStateComponent<PersistentState> { // has to be public (for IDEA configuration access)
 
     private boolean enhancerEnabled = true;
 
+    @XCollection(elementTypes = String.class)
     private Collection<String> metaDataExtensions = new ArrayList<String>(Arrays.asList("jdo", "orm"));
 
     /**
@@ -28,13 +36,19 @@ public class PersistentState { // has to be public (for IDEA configuration acces
     private boolean enforcePropertyRestrictions = true;
     private boolean tmpClassLoader = true;
 
+    @XCollection(elementTypes = String.class)
     private Collection<String> enabledModules = new ArrayList<String>();
 
+    @XCollection(elementTypes = String.class)
     private Collection<String> enabledFiles = new ArrayList<String>();
 
     private String api = "JPA";
 
     private String enhancerSupport = "OPENJPA";
+
+    static PersistentState getInstance(Project project) {
+        return project.getService(PersistentState.class);
+    }
 
     public boolean isEnhancerEnabled() {
         return this.enhancerEnabled;
@@ -44,12 +58,10 @@ public class PersistentState { // has to be public (for IDEA configuration acces
         this.enhancerEnabled = enhancerEnabled;
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public Collection<String> getMetaDataExtensions() {
         return new LinkedHashSet<String>(this.metaDataExtensions);
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public void setMetaDataExtensions(final Collection<String> metaDataExtensions) {
         this.metaDataExtensions = new LinkedHashSet<String>(metaDataExtensions);
     }
@@ -70,22 +82,18 @@ public class PersistentState { // has to be public (for IDEA configuration acces
         this.includeTestClasses = includeTestClasses;
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public Collection<String> getEnabledModules() {
         return new LinkedHashSet<String>(this.enabledModules);
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public void setEnabledModules(final Collection<String> enabledModules) {
         this.enabledModules = new LinkedHashSet<String>(enabledModules);
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public Collection<String> getEnabledFiles() {
         return new LinkedHashSet<String>(this.enabledFiles);
     }
 
-    @AbstractCollection(elementTypes = String.class)
     public void setEnabledFiles(final Collection<String> enabledFiles) {
         this.enabledFiles = new LinkedHashSet<String>(enabledFiles);
     }
@@ -131,59 +139,13 @@ public class PersistentState { // has to be public (for IDEA configuration acces
         this.tmpClassLoader = tmpClassLoader;
     }
 
-    /**
-     * Copy method used to update persistent state with plugin's internal state.
-     *
-     * @param state plugin's internal state
-     * @return persistent copy of plugin's internal state
-     */
-    @SuppressWarnings({"FeatureEnvy", "ChainedMethodCall"})
-    public PersistentState copyFrom(final State state) {
-        this.enhancerEnabled = state.isEnhancerEnabled();
-
-        if (this.metaDataExtensions == null) {
-            this.metaDataExtensions = new ArrayList<String>();
-        } else {
-            this.metaDataExtensions.clear();
-        }
-        this.metaDataExtensions.addAll(state.getMetaDataExtensions());
-
-        this.includeTestClasses = state.isIncludeTestClasses();
-        this.addDefaultConstructor = state.isAddDefaultConstructor();
-        this.enforcePropertyRestrictions = state.isEnforcePropertyRestrictions();
-        this.tmpClassLoader = state.isTmpClassLoader();
-
-        if (this.enabledModules == null) {
-            this.enabledModules = new ArrayList<String>();
-        } else {
-            this.enabledModules.clear();
-        }
-        this.enabledModules.addAll(state.getEnabledModules());
-        if (this.enabledFiles == null) {
-            this.enabledFiles = new ArrayList<String>();
-        } else {
-            this.enabledFiles.clear();
-        }
-        this.enabledFiles.addAll(state.getEnabledFiles());
-
-        final EnhancerSupport configuredEnhancerSupport = state.getEnhancerSupport();
-        final EnhancerSupport usedEnhancerSupport;
-        if (configuredEnhancerSupport == null) {
-            usedEnhancerSupport = state.getEnhancerSupportRegistry().getDefaultEnhancerSupport();
-            this.enhancerSupport = usedEnhancerSupport.getId();
-        } else {
-            usedEnhancerSupport = configuredEnhancerSupport;
-            this.enhancerSupport = configuredEnhancerSupport.getId();
-        }
-
-        final PersistenceApi myApi = state.getApi();
-        if (myApi == null || !usedEnhancerSupport.isSupported(myApi)) {
-            this.api = usedEnhancerSupport.getDefaultPersistenceApi().name();
-        } else {
-            this.api = myApi.name();
-        }
-
+    @Override
+    public PersistentState getState() {
         return this;
     }
 
+    @Override
+    public void loadState(@NotNull PersistentState state) {
+        XmlSerializerUtil.copyBean(state, this);
+    }
 }
