+++ src/main/java/org/openjpa/ide/idea/integration/EnhancerSupport.java
@@ -4,11 +4,13 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
 import com.intellij.openapi.compiler.CompileContext;
+import com.intellij.openapi.extensions.ExtensionPointName;
 import com.intellij.openapi.module.Module;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.openjpa.ide.idea.PersistenceApi;
+import org.openjpa.ide.idea.ProjectComponent;
 
 /**
  * Interface to implement for every new enhancer to support.<br/>
@@ -16,7 +18,7 @@ import org.openjpa.ide.idea.PersistenceApi;
  */
 public interface EnhancerSupport {
 
-    public static final String EXTENSION_POINT_NAME = "OpenJpaIntegration.openjpaEnhancerExtension";
+    ExtensionPointName<EnhancerSupport> EXTENSION_POINT_NAME = ExtensionPointName.create("OpenJpaIntegration.openjpaEnhancerExtension");
 
     /**
      * The name to display in the configuration dialog enhancer support drop-down.
